\name{data.anscombe {desk}}
\docType{data}
\alias{data.anscombe}
\title{
Anscombe's Four Data Sets
}
\description{
This data set comprises four individual x-y-data sets which have the same statistical properties (mean, variance, correlation, regression line, etc.), yet are quite different.
}
\usage{data.anscombe}
\format{
  A data frame of 4 data sets, each with 11 observations of the two variables x and y.
  \describe{
    \item{x1 to x4}{x-variables of the four data sets.}
    \item{y1 to y4}{y-variables of the four data sets.}
  }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 3) these data are used to illustrate the simple regression model.
}
\source{
Anscombe, F. J. (1973) Graphs in Statistical Analysis. American Statistician, 27(1), 17-21.
}
\references{
Tufte, E. R. (1989) \emph{The Visual Display of Quantitative Information}, 13-14. Graphics Press.

Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler	(\url{http://www.oekonometrie-lernen.de}).
		}

\keyword{datasets}
