\name{data.auto {desk}}
\alias{data.auto}
\docType{data}
\title{
Prices and Qualitative Characteristics of US-Cars
}
\description{
This is a data set on the prices and qualitative characteristics of US-cars sold in 1979.}
\usage{data.auto}
\format{
  A data frame with 52 observations on the following nine variables:
  \describe{
    \item{make}{Make and model.}
    \item{price}{Price (in dollar).}
    \item{mpgall}{Mileage (miles per gallon).}
    \item{headroom}{Headroom (in inch).}
    \item{trunk}{Trunk Space (in cubic foot).}
    \item{weight}{Weight (in pound).}
    \item{length}{Length (in inch).}
    \item{turn}{Turn circle (in foot).}
    \item{displacement}{Displacement (in cubic inch).}
    }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 13) these data are used to illustrate the selection process of exogenous variables.
}
\source{
Chambers, J.M, W.S. Cleveland, B. Kleiner, P.A. Tukey (1983) \emph{Graphical Methods for Data Analysis}, Wadsworth International Group.
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
		}

\keyword{datasets}
