\name{data.ballb {desk}}
\alias{data.ballb}
\docType{data}
\title{
Defective Ball Bearings
}
\description{
This is a data set on the percentage of rejects in the production of ball bearings.
}
\usage{data.ballb}
\format{
  A data frame with seven observations on the following two variables:
  \describe{
    \item{defbb}{Share of defective ball bearings (per thousand).}
    \item{nshifts}{Number of shifts between two maintenances.}
  }
}
\details{
In Auer, L. von (2016, Chap. 16) and Auer, L. von, S. Hoffmann (2017, Chap. 16) these hypothetical data are used to illustrate the consequences of error terms with an expected value deviating from zero.
}
\source{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
