\name{data.burglary {desk}}
\alias{data.burglary}
\docType{data}
\title{
Burglaries and Power Blackouts
}
\description{
This is a dataset on the monthly number of burglaries and the number of power blackouts in a small town.
}
\usage{data.burglary}
\format{
  A data frame with 12 observations on the following three variables:
  \describe{
    \item{month}{Month.}
    \item{burglary}{Number of burglaries.}
    \item{blackout}{Number of power blackouts.}
  }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 15) these hypothetical data are used to illustrate the consequences of a structural break.
}
\source{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
  	}

\keyword{datasets}
