\name{data.cars {desk}}
\alias{data.cars}
\docType{data}
\title{
Speed and Stopping Distances of Cars
}
\description{
The data give the speed of cars and the distances taken to stop. The data were recorded in the 1920s.
}
\usage{data.cars}
\format{
  A data frame of 50 observations with the following two variables:
  \describe{
    \item{speed}{Speed (in miles per hour)}
    \item{dist}{Stopping distance (in foot)}
  }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chaps. 5, 6, 7, 16) the data are used to illustrate the simple regression model and the consequences of truncated data.
}
\source{
R package datasets (object cars)

Ezekiel, M. (1930) \emph{Methods of Correlation Analysis}, Wiley.
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler  (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
