\name{data.cobbdoug {desk}}
\alias{data.cobbdoug}
\docType{data}
\title{
Cobb-Douglas Production Function
}
\description{
This is a data set on a production process that can be formalized by a Cobb-Douglas production function.
}
\usage{data.cobbdoug}
\format{
  A data frame with 100 observations on the following three variables:
  \describe{
    \item{output}{Production output.}
    \item{labor}{Input of labor.}
    \item{capital}{Input of capital.}
  }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 14) these hypothetical data are used to illustrate the functional specification of a regression model.
}
\source{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
