\name{data.comp {desk}}
\alias{data.comp}
\docType{data}
\title{
Monthly Rentals and Qualitative Characteristics of Computers
}
\description{
This is a data set on the monthly rentals of computers of different quality during the 1960s.
}
\usage{data.comp}
\format{
  A data frame with 34 observations on the following four variables:
  \describe{
    \item{rent}{Monthly rental (in dollar).}
    \item{mem}{Memory capacity computed from three different computer characteristics.}
    \item{access}{Average time required to access information from memory.}
    \item{mult}{Average time required to obtain and complete multiplication instruction.}
  }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chaps. 13 and 14) these hypothetical data are used to illustrate the specification of a multivariate regression model.
}
\source{
Berndt, E.R. (1990) \emph{The Practice of Econometrics: Classic and Contemporary}, Addison-Wesley, 136-142.
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
