\name{data.fertilizer {desk}}
\docType{data}
\alias{data.fertilizer}
\title{
Fertilizer in the Cultivation of Barley
}
\description{
This is a data set on the use of fertilizers (phospate and nitrogen) in the cultivation of barley.
}
\usage{data.fertilizer}
\format{
  A data frame with 30 observations on the following three variables:
  \describe{
    \item{phos}{Amount of phosphate (in kg per hectare).}
    \item{nit}{Amount of nitrogen (in kg per hectare).}
    \item{barley}{Barley crop yield (in units of 100 kg per hectare).}
  }
}
\details{
In Auer, L von (2016, Chap. 9) and Auer, L. von, S. Hoffmann (2017, Chap. 9) these hypothetical data are used to illustrate the estimation of a multivariate linear regression model.
}
\source{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
