\name{data.filter {desk}}
\alias{data.filter}
\docType{data}
\title{
Water Filter Sales
}
\description{
This is a data set on the prices and sales figures of water filters (in 1000 pcs.).
}
\usage{data.filter}
\format{
A data frame with 24 observations on the following two variables:
  \describe{
    \item{sales}{Monthly water filter sales (in 1000 pcs.).}
    \item{price}{price (in Euro).}
  }
}
\details{
In Auer, L von (2016, Chap. 18) and Auer, L. von, S.Hoffmann (2017, Chap. 18) these hypothetical data are used to illustrate the consequences of autocorrelated error terms.
}
\source{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
