\name{data.icecream {desk}}
\alias{data.icecream}
\docType{data}
\title{
Sales of Ice Cream
}
\description{
This data set is on the daily revenues from selling ice cream and the daily average temperature in some town on a sample of 35 working days.
}
\usage{data.icecream}
\format{
  A data frame with 35 observations on the following two variables:
  \describe{
    \item{revenue}{Revenues (in Euro).}
    \item{temp}{Temperature (in degree Celsius).}
  }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 7) these hypothetical data are used to illustrate the estimation of the simple linear regression model.
}
\source{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
}
\keyword{datasets}
