\name{data.income {desk}}
\alias{data.income}
\docType{data}
\title{
Income Per Capita
}
\description{
This data set describes major macroeconomic variables determining the differences in per capita income of 75 countries in 1985.
}
\usage{data.income}
\format{
  A data frame with 75 observations on the following three variables:
  \describe{
    \item{loginc}{Logarithmic per capita income.}
    \item{logsave}{Logarithmic savings rate.}
    \item{logsum}{Logarithmic sum of population growth rate, technical progress and capital depreciation.}
  }
}
\details{
In Auer, L von (2016, Chap. 19) and Auer, L. von, S. Hoffmann (2017, Chap. 19) these hypothetical data are used to illustrate the detection and consequences of error terms that are not normally distributed.
}
\source{
Mankiw, Romer and Weil (1992), A Contribution to the Empirics of Economic Growth, Quarterly Journal of Economics, 107, 407-437.
}
\references{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).

Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
