\name{data.insurance {desk}}
\alias{data.insurance}
\docType{data}
\title{
Sales of Insurance Contracts
}
\description{
This is a data set on the ability and success of salespersons in selling insurance contracts.
}
\usage{data.insurance}
\format{
  A data frame with 30 observations on the following four variables:
  \describe{
    \item{contr}{Number of insurance contracts currently sold by the salesperson.}
    \item{score}{Score of salesperson in assessment center.}
    \item{contrprev}{Number of insurance contracts sold period by the salesperson in the previous.}
    \item{ability}{Salesperson's true ability to sell insurance contracts.}
  }
}
\details{
In Auer, L von (2016, Chap. 20) and Auer, L. von, S. Hoffmann (2017, Chap. 20) these hypothetical data illustrate the use of two stage least squares estimation with an instrumental variable.
}
\source{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
