\name{data.iv {desk}}
\alias{data.iv}
\docType{data}
\title{
Instrumental Variables
}
\description{
This data set is on the use of instrumental variables.
}
\usage{data.iv}
\format{
  A data frame with 8 observations on the following five variables:
  \describe{
    \item{y}{Endogenous variable.}
    \item{x1}{First exogenous variable.}
    \item{x2}{Second exogenous variable.}
    \item{z1}{First instrumental variable.}
    \item{z2}{Second instrumental variable.}
  }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 20) these hypothetical data are used to illustrate the use of two stage least squares estimation with instrumental variables.
}
\source{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\references{

}

\keyword{datasets}
