\name{data.kmenta}
\alias{data.kmenta}
\docType{data}
\title{
Partly Artificial Data on the U. S. Economy
}
\description{
These are partly contrived data from Kmenta (1986), constructed to illustrate estimation of a simultaneous-equation model. 
}
\usage{data.kmenta}
\format{
  A data frame of 20 observations with the following 5 variables:
  \describe{
    \item{consump}{food consumption per capita.}
    \item{price}{ratio of food prices to general consumer prices.}
    \item{income}{disposable income in constant dollars.}
    \item{farmPrice}{ratio of preceding year's prices received by farmers to general consumer prices.}
    \item{trend}{time trend in years.}
  }
}
\details{
The exogenous variables income, farmPrice, and trend are based on real data; the endogenous variables price and consump were generated by simulation. 
}
\source{
R package \code{systemfit}
}
\references{
Kmenta, J. (1986) \emph{Elements of Econometrics}, 2nd Edition, Macmillan, New York.
}

\keyword{datasets}
