\name{data.lifesat {desk}}
\alias{data.lifesat}
\docType{data}
\title{
Life Satisfaction
}
\description{
A data set describing the life satisfaction and per capita income in 40 countries in 2010.
}
\usage{data.lifesat}
\format{
  A data frame of 40 observations with the following three variables:
  \describe{
    \item{country}{Country name.}
    \item{income}{Country's per capita income (in dollar).}
    \item{lsat}{Index of country's average life satisfaction.}
  }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 3) these hypothetical data are used to illustrate the use of the simple linear regression model.
}
\source{
World Value Survey
\url{http://www.worldvaluessurvey.org}
}

\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
