\name{data.milk {desk}}
\alias{data.milk}
\docType{data}
\title{
Milk Production
}
\description{
This is a data set on the use of concentrated feed for cows and their milk output.
}
\usage{data.milk}
\format{
  A data frame with 12 observations on the following two variables:
  \describe{
    \item{feed}{Concentrated feed given to the cow (in units of 50kg per year).}
    \item{milk}{Milk output of the cow (in litres per year).}
  }
}
\details{
In Auer, L von (2016, Chap. 14) and Auer, L. von, S. Hoffmann (2017, Chap. 14) these hypothetical data are used to illustrate the functional specification of a regression model.
}
\source{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
