\name{data.pharma {desk}}
\alias{data.pharma}
\docType{data}
\title{
Pharmaceutical Advertisements
}
\description{
This is a data set on quarterly commercials data of a pharmaceutical company.
}
\usage{data.pharma}
\format{
  A data frame with 24 quarterly observations on the following four variables:
  \describe{
    \item{sales}{Sales of pharmaceutical product (in units of 100g).}
    \item{ads}{Number of advertisements (in double pages).}
    \item{price}{Price of pharmaceutical product (in euro per 100g).}
    \item{adsprice}{Price of advertisements (in units of 1000 euro per double page).}
  }
}
\details{
In Auer, L von (2016, Chap. 23) and Auer, L. von, S. Hoffmann (2017, Chap. 23) these hypothetical data are used to illustrate the estimation of simultaneous equation econometric models.
}
\source{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
