\name{data.printer {desk}}
\alias{data.printer}
\docType{data}
\title{
Prices and Qualitative Characteristics of Laser Printers
}
\description{
This is a data set on the prices and qualitative characteristics of laser printers from 1992 to 2001.
}
\usage{data.printer}
\format{
  A data frame with 44 observations on the following five variables:
  \describe{
    \item{price}{Price of the printer (in euro).}
    \item{speed}{Printer's speed (in pages per minute).}
    \item{size}{Printer's size (in cubic decimeter).}
    \item{mcost}{Maintenance costs of printer (in cent per page).}
    \item{tdiff}{Time difference between the printer's observation and the data set's first observed laser printer (in month).}
  }
}
\details{
In Auer, L von (2016, Chap. 21) and Auer, L. von, S. Hoffmann (2017, Chap. 21) these hypothetical data are used to illustrate the consequences of multicollinear exogenous variables.
}
\source{
Data from computer magazin c't (February 1992 to August 2001).

Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
