\name{data.rent {desk}}
\alias{data.rent}
\docType{data}
\title{
Average Basic Rent in City Districts
}
\description{
This is a data set on twelve districts of a city. The data describe the district's distance to the city center and the average basic rent (it excludes additional costs).
}
\usage{data.rent}
\format{
  A data frame with 12 observations on the following two variables:
  \describe{
    \item{dist}{Distance between district and city center (in km).}
    \item{rent}{District's basic rent (in euro per square meter).}
  }
}
\details{
In Auer, L von (2016, Chap. 17) and Auer, L. von, S. Hoffmann (2017, Chap. 17) these hypothetical data are used to illustrate the consequences of heteroskedastic error terms.
}
\source{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
