\name{data.sick {desk}}
\alias{data.sick}
\docType{data}
\title{
Sick Leave and Unemployment
}
\description{
This is a data set on the unemployment rates and the sick leave in Germany in the years 1992 to 2014.
}
\usage{data.sick}
\format{
  A data frame with 23 observations on the following three variables:
  \describe{
    \item{year}{Year.}
    \item{jobless}{Average unemployment rate during that year (in percent).}
    \item{sick}{Average of employees' sick leave during that year (in percent).}
  }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 18) these data are used to illustrate the consequences of autocorrelated error terms.
}
\source{
Federal Statistical Office of Germany \url{https://www.destatis.de/DE/ZahlenFakten/Indikatoren/QualitaetArbeit/QualitaetDerArbeit.html?cms_gtp=318944_slot\%3D2&https=1}
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
