\name{data.software {desk}}
\alias{data.software}
\docType{data}
\title{
Employment Data of a Software Company
}
\description{
This is a data set on business data of a software company covering 36 consecutive months.
}
\usage{data.software}
\format{
  A data frame with 36 observations on the following two variables:
  \describe{
    \item{empl}{Number of employees in the observed month.}
    \item{orders}{Number of new orders during the observed month.}
  }
}
\details{
In Auer, L von (2016, Chap. 22) and Auer, L. von, S. Hoffmann (2017, Chap. 22) these hypothetical data are used to illustrate the estimation of dynamic regression models.
}
\source{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
