\name{data.tip {desk}}
\alias{data.tip}
\docType{data}
\title{
Tip Data in a Restaurant
}
\description{
This is a data set on the bills and the corresponding tips given in a restaurant.
}
\usage{data.tip}
\format{
  A data frame with three observations on the following two variables:
  \describe{
    \item{x}{The guest's bill (in euro).}
    \item{y}{The tip given to the waiter/waitress (in euro).}
  }
}
\details{
In Auer, L von (2016, Chap. 17) and Auer, L. von, S. Hoffmann (2017, Chap. 17) these hypothetical data provide a minimal dataset for estimating a simple linear regression model.
}
\source{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\keyword{datasets}