\name{data.unempl {desk}}
\alias{data.unempl}
\docType{data}
\title{
German Economic Growth and Unemployment Rates
}
\description{
This is a data set on German economic growth and unemployment rates from 1992 to 2014.
}
\usage{data.unempl}
\format{
  A data frame with 23 observations on the following three variables:
  \describe{
    \item{year}{Year.}
    \item{unempl}{Change in German unemployment rate (in percentage points).}
    \item{gdp}{Change in German gross domestic product (in percentage points).}
  }
}
\details{
In Auer, L von (2016, Chap. 15) and Auer, L. von, S. Hoffmann (2017, Chap. 15) these yearly data are used to illustrate the estimation of regression models that exhibit a structural break.
}
\source{
Online platform "Genesis" of the Federal Statistical Office of Germany
\url{http://www.destatis.de}

Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
