\name{data.wage {desk}}
\alias{data.wage}
\docType{data}
\title{
Wage Data in a Company
}
\description{
This is a data set on the wage structure in a company.
}
\usage{data.wage}
\format{
  A data frame with 20 observations on the following four variables:
  \describe{
    \item{wage}{Employee's monthly wage (in euro).}
    \item{educ}{Employee's extra education beyond the basic schooling degree (in years).}
    \item{age}{Employee's age (in years).}
    \item{empl}{Employee's time of employment in the company (in years).}
  }
}
\details{
In Auer, L von (2016, Chap. 13) and Auer, L. von, S. Hoffmann (2017, Chap. 13) these hypothetical data are used to illustrate the selection of the relevant exogenous variables.
}
\source{
Auer, L. von (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
