\name{data.windscreen {desk}}
\alias{data.windscreen}
\docType{data}
\title{
Efficiency of a Car Glass Service Company
}
\description{
This is a data set on the business statistics of 248 branches of a car glass service company in 2015.
}
\usage{data.windscreen}
\format{
  A data frame with 248 observations on the following eight variables:
  \describe{
    \item{screen}{Number of windscreen replacements in the branch.}
    \item{foreman}{Foremen employed in the branch.}
    \item{assist}{Assistants employed in the branch.}
    \item{f_wage}{Foremen's average wage in the branch.}
    \item{a_wage}{Assistants' average wage in the branch.}
    \item{f_age}{Foremen's average age in the branch.}
    \item{a_age}{Assistants' average age in the branch.}
    \item{capital}{Total value of machines used for windscreen replacement in the branch (in euro).}
  }
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 20) these data illustrate the use of two stage least squares estimation with instrumental variables.
}
\source{
Auer, L. von, S. Hoffmann (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{http://www.oekonometrie-lernen.de}).
}
\references{
}
\keyword{datasets}
