\name{ols.infocrit}

\alias{ols.infocrit}

\title{Calculate common information criteria.}

\description{
Calculates three common information criteria of models estimated by \code{ols()}.
}

\usage{ols.infocrit(mod, which = "all", scaled = FALSE)}

\arguments{
  \item{mod}{Linear model object generated by ols()}

  \item{which}{String value specifying the type of criterion: "aic" (Akaike Information Criterion), "sic" (Schwarz Information Criterion), or
  "pc", (Prognostic Criterion), optional, if ommitted then all criteria are returned ("all").}
  
  \item{scaled}{logical value which indicates whether criteria should be scaled by the number of observations T}
}

\value{Dataframe of AIC, SIC, and PC values}

\examples{
  wage.est <- ols(wage ~ educ + age, data = data.wage)
  ols.infocrit(wage.est) # Return all criteria unscaled
  ols.infocrit(wage.est, scaled = TRUE) # Return all criteria scaled
  ols.infocrit(wage.est, which = "pc") # Return Prognostic Criterion unscaled
}

\keyword{Akaike Information Criterion, Schwarz Criterion, Prognostic Criterion}