\name{plot.desk}

\alias{plot.desk}

\title{Simplified plotting of regression- and test-results.}
\description{This function implements an S3 method for plotting regression- and test-results generated by functions of the desk package. Used for internal purposes.
}

\usage{
\method{plot}{desk}(x, ...)
}

\arguments{
  \item{x}{Object of class desk to be plotted.}
  \item{...}{Any argument that \code{plot()} accepts.}
}

\value{None}

\examples{
# see the help page of the commands, respectively
}

\keyword{plot, regression results, test results}